+++ game/resource/addon_schinese.txt
@@ -161,11 +161,11 @@
 		//			Lottery Item/Ability
 		//
 		///////////////////////////////////////////////////
-		"lottery_header"										"选择一个道具！"
+		"lottery_title"											"选择一个物品"
 
-		"lottery_item_refresh"									"点击刷新道具"
-		"lottery_item_refresh_used"								"道具已刷新"
-		"lottery_item_refresh_not_member"						"会员每日可刷新一次道具，点击开通会员"
+		"lottery_item_refresh"									"点击刷新物品"
+		"lottery_item_refresh_used"								"物品已刷新"
+		"lottery_item_refresh_not_member"						"会员每日可刷新一次物品，点击开通会员"
 		"lottery_ability_refresh"								"点击刷新技能"
 		"lottery_ability_refresh_used"							"技能已刷新"
 		"lottery_ability_refresh_not_member"					"会员每日可刷新一次技能，点击开通会员"
@@ -876,7 +876,7 @@
 		"DOTA_Tooltip_ability_artoria_mana_burst_duration"		"减速持续时间："
 
 		"DOTA_Tooltip_Ability_artoria_excalibur"				"<font color='#FFE56E'>誓约胜利之剑</font>"
-		"DOTA_Tooltip_ability_artoria_excalibur_description"	"阿尔托莉雅将魔力转化为'光'，通过压缩、加速增加运动能量，使神灵级魔术的行使成为可能。对剑光路径上的每个敌人造成伤害。\n\n<font color='#FF3399' >道具EX咖喱棒强化：誓约胜利之剑。</font>"
+		"DOTA_Tooltip_ability_artoria_excalibur_description"	"阿尔托莉雅将魔力转化为'光'，通过压缩、加速增加运动能量，使神灵级魔术的行使成为可能。对剑光路径上的每个敌人造成伤害。\n\n<font color='#FF3399' >物品EX咖喱棒强化：誓约胜利之剑。</font>"
 		"DOTA_Tooltip_ability_artoria_excalibur_Lore"	"孤傲之王的念想灼耀星夜。"
 		"DOTA_Tooltip_ability_artoria_excalibur_cast_delay"		"施法延迟："
 		"DOTA_Tooltip_ability_artoria_excalibur_width"		"宽度："
@@ -1496,7 +1496,7 @@
 
 		// item_blue_fantasy 苍蓝幻想
 		"DOTA_Tooltip_ability_item_blue_fantasy"									"苍蓝幻想"
-		"DOTA_Tooltip_ability_item_blue_fantasy_description"						"<h1><font color='#FFFF00'>主动：毁灭终曲</font></h1>施加持续%mute_duration%秒的<font color='#f15b47'>被动破坏，装备禁用，闪避失效</font>的负面效果。<br><br>持续驱散目标身上的增益效果，降低%slow_pct%%%移速，%attack_slow%攻速，生命恢复、治疗、吸血和技能吸血效果降低%hp_regen_reduction%%%。<br><br>目标每秒受到自身最大生命值<font color='#ffcc33'>%max_hp_dmg_pct%%%</font>的纯粹伤害。\n<h1>被动：蚕食进行曲</h1>每次攻击燃烧目标 %mana_base% 点魔法并造成等量的物理伤害。(多件道具不叠加)"
+		"DOTA_Tooltip_ability_item_blue_fantasy_description"						"<h1><font color='#FFFF00'>主动：毁灭终曲</font></h1>施加持续%mute_duration%秒的<font color='#f15b47'>被动破坏，装备禁用，闪避失效</font>的负面效果。<br><br>持续驱散目标身上的增益效果，降低%slow_pct%%%移速，%attack_slow%攻速，生命恢复、治疗、吸血和技能吸血效果降低%hp_regen_reduction%%%。<br><br>目标每秒受到自身最大生命值<font color='#ffcc33'>%max_hp_dmg_pct%%%</font>的纯粹伤害。\n<h1>被动：蚕食进行曲</h1>每次攻击燃烧目标 %mana_base% 点魔法并造成等量的物理伤害。(多件物品不叠加)"
 		"DOTA_Tooltip_ability_item_blue_fantasy_int"								"+智力"
 		"DOTA_Tooltip_ability_item_blue_fantasy_agi"								"+敏捷"
 		"DOTA_Tooltip_ability_item_blue_fantasy_bonus_damage"						"+攻击力"
@@ -1589,7 +1589,7 @@
 		// 苍洋魔珠
 		"DOTA_Tooltip_Ability_item_orb_of_the_brine"						"苍洋魔珠"
 		"DOTA_Tooltip_Ability_item_orb_of_the_brine_Lore"						"要不是这颗尊崇的魔珠，教团在石鱼门一役就全军覆没。"
-		"DOTA_Tooltip_Ability_item_orb_of_the_brine_Description"			"<h1>主动：苍洋泡沫</h1> 使一个友方单位进入具有无敌效果的疗伤泡泡，移动速度降低至%bubble_move_speed%，无法攻击释放技能和道具持续%bubble_duration%秒。<br/>回复%bubble_total_heal_tip%点生命，%bubble_total_mana_tip%点魔法。\n\n<h1>被动：苍洋福佑</h1>自身提供的治疗效果提升%heal_increase%%%。"
+		"DOTA_Tooltip_Ability_item_orb_of_the_brine_Description"			"<h1>主动：苍洋泡沫</h1> 使一个友方单位进入具有无敌效果的疗伤泡泡，移动速度降低至%bubble_move_speed%，无法攻击释放技能和物品持续%bubble_duration%秒。<br/>回复%bubble_total_heal_tip%点生命，%bubble_total_mana_tip%点魔法。\n\n<h1>被动：苍洋福佑</h1>自身提供的治疗效果提升%heal_increase%%%。"
 		"DOTA_Tooltip_Ability_item_orb_of_the_brine_bonus_all_stats"		"+$all"
 		"DOTA_Tooltip_Ability_item_orb_of_the_brine_bonus_health"			"+$health"
 		"DOTA_Tooltip_Ability_item_orb_of_the_brine_bonus_mana"				"+$mana"
